refactor(frontend): extract shared DataAgent widget loader util
Deduplicate the identical loadDataAgentWidgetScript function and its
constants from Layout.vue and IntelligentQueryRemoteEmbed.vue into a
single shared util at src/utils/dataagentWidgetLoader.js.
